+++ src/Enums/ResolutionKind.php
@@ -6,9 +6,7 @@ namespace Relaticle\CustomFields\Enums;
 
 enum ResolutionKind: string
 {
-    case Form = 'form';
     case Infolist = 'infolist';
     case Table = 'table';
     case Exporter = 'exporter';
-    case Importer = 'importer';
 }