+++ src/Models/CustomField.php
@@ -12,6 +12,7 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 use Override;
 use Relaticle\CustomFields\CustomFields;
@@ -210,9 +211,7 @@ class CustomField extends Model
 
     public function setting(string $key, mixed $default = null): mixed
     {
-        $additional = $this->settings->additional;
-
-        return array_key_exists($key, $additional) ? $additional[$key] : $default;
+        return Arr::get($this->settings->additional, $key, $default);
     }
 
     public function getCurrencySettings(): CurrencyFieldSettingsData