+++ src/Models/CustomFieldSection.php
@@ -10,6 +10,7 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Support\Arr;
 use Relaticle\CustomFields\CustomFields;
 use Relaticle\CustomFields\Data\CustomFieldSectionSettingsData;
 use Relaticle\CustomFields\Database\Factories\CustomFieldSectionFactory;
@@ -90,9 +91,7 @@ class CustomFieldSection extends Model
 
     public function setting(string $key, mixed $default = null): mixed
     {
-        $extra = $this->settings->extra;
-
-        return array_key_exists($key, $extra) ? $extra[$key] : $default;
+        return Arr::get($this->settings->extra, $key, $default);
     }
 
     /**